Backend/requirements.txt
requests~=2.32.3
fastapi~=0.111.0
SQLAlchemy~=2.0.31
pydantic~=2.8.2
python-dotenv~=1.0.1
mysql-connector-python